implemented sedd backup logic
