+++ android/src/main/java/io/parity/signer/models/SignerDataModel.kt
@@ -15,7 +15,6 @@ import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.security.crypto.EncryptedSharedPreferences
 import androidx.security.crypto.MasterKey
-import io.parity.signer.components.Authentication
 import io.parity.signer.dependencyGraph.getDbNameFromContext
 import io.parity.signer.ui.navigationselectors.OnBoardingState
 import io.parity.signer.uniffi.*
