+++ android/src/main/res/values/strings.xml
@@ -52,4 +52,5 @@
 	<string name="auth_error_status_unknown">Authentication system failure</string>
 	<string name="subtitle_secret_recovery_phrase">Secret Recovery Phrase</string>
 	<string name="subtitle_derived_keys">Derived Keys</string>
+    <string name="seed_backup_autohide_title">Backup info will be hidden in</string>
 </resources>
